pyproject.toml
[project]
name = "code-review-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Aadi Burande",email = "aadiburande@gmail.com"}
]
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langchain (>=0.2.0)",
    "langgraph (>=0.1.0)",
    "openai (>=1.30.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pylint (>=4.0.5,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "radon (>=6.0.1,<7.0.0)",
    "semgrep (>=1.154.0,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
